rename ressources to resources
