+++ rio_viz/resources/__init__.py
@@ -1,0 +1,1 @@
+"""titiler.resources."""
