--- rio_viz/ressources/__init__.py
@@ -1,1 +1,0 @@
-"""titiler ressources."""
